Make file structure uniform for both v1alpha1 and v1alpha2
