+++ apis/api/v1alpha1/astartevoyageringress_types.go
@@ -155,7 +155,3 @@ type AstarteVoyagerIngressStatus struct {
 	// Important: Run "operator-sdk generate k8s" to regenerate code after modifying this file
 	// Add custom validation using kubebuilder tags: https://book-v1.book.kubebuilder.io/beyond_basics/generating_crd.html
 }
-
-func init() {
-	SchemeBuilder.Register(&AstarteVoyagerIngress{}, &AstarteVoyagerIngressList{})
-}
