--- apis/api/v1alpha2/groupversion_info.go
@@ -1,33 +1,0 @@
-/*
-  This file is part of Astarte.
-
-  Copyright 2020 Ispirata Srl
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-
-// Package v1alpha2 contains API Schema definitions for the api v1alpha2 API group
-// +kubebuilder:object:generate=true
-// +groupName=api.astarte-platform.org
-package v1alpha2
-
-var (
-// GroupVersion is group version used to register these objects
-// GroupVersion = schema.GroupVersion{Group: "api.astarte-platform.org", Version: "v1alpha2"}
-
-// SchemeBuilder is used to add go types to the GroupVersionKind scheme
-// SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
-
-// AddToScheme adds the types in this group-version to the given scheme.
-// AddToScheme = SchemeBuilder.AddToScheme
-)
